third_party/WebKit/LayoutTests/fast/frames/seamless/seamless-sandbox-srcdoc-expected.txt

Test that iframes with srcdoc contents can still be seamless.
PASS successfullyParsed is true

TEST COMPLETE
PASS iframe.seamless is true
PASS window.getComputedStyle(iframe).width is "200px"
PASS window.getComputedStyle(iframe).height is "100px"
